separate fields for json and datastore.
a result of map[string]Connections leads to a wrong generation for the
swagger generated go sources. so we split the result for the client in a
separate field which is a simple array of connections. the internal
storage is still a map, so we have an efficient storage/lookup.