move metal package to be internal.
All types of metal-api should not be importable from other go programs.
Instead the openapi-spec should be used to generate a client. To have
better types in the spec, the request-parameters are now also in the
new internal metal package and exported by this. So the spec should
now have consistent naming for the datatypes.