Do not allow individual capacity counts to be higher than total + refactor.
The counter-intuitive sum was caused by a machine that had issues and that was allocated at the same time, which was leading to the sum of the individual counts being higher than total.
The count can still be lower than the total count because there can be machine that are for example "preparing" at the moment. They are neither allocated nor free nor faulty.