Set VRF at switches before removing machine allocation.
This targets to mitigate the case that when machine free fails in between (after removing allocation from machine but before reconfiguring the switches). This does not resolve the issue of properly freeing a machine, but ensures that a machine cannot reboot back into the prior OS and being able to connect to the internet again. When the machine can reach the internet again, user services on the machine can start running again as if nothing happened, which is dangerous.