Free machine async (#66)
* implement async free machine

* change test data

* first delete the IP in our store, then release it.

by first deleting the IP in our store, the IP is still allocated in the
IPAM and cannot be used by another if some error occurs while deleting.
It will only be released, when it is successfully deleted or it does not
exist any more.

* loglevel should be error, when an error occured.

we should not log to info when an error happened which should not
happen. instead we must log with level "Error" so we later can crawl the
logfile for errors.

* one rest file for testing.

the allocation of the machine uses the network-id of the registered
network which has to be called before.

* implement new api for async functions

* use latest metal-lib

* repair linter errors

* describe intention with a comment

* use my own logger instead of the logger of the embedded rethinkstore

* return error instead of panic in IP Webservice

* return error instead of panic in machine Webservice

* move validations for ip release to the caller.

the function for ip releasing must only be called after the caller
validated if the release is ok. the function itself is called in
different situations, so there is no single validation for all cases.