Add data store connect opts.
The intention is not to run database initialization or database user demotion when it's not required.

Running database initialization can be problematic when scenarios like the following occur:

- The rethinkdb does not come during the deployment for some reason
- Commands from cron jobs like evaluateMachineLiveliness, resurrectDeadMachines, ... queue up in Kubernetes
- When the database comes up all the pods start execution at the same time
- If there was a change in the tables, it can happen that tables are created multiple times -> Rethinkdb is broken

Generally, we should configure the deployment such that only the init-db pre-deployment-hook initializes the database because the hook is guaranteed to only run one at a time.